+++ packages/next/src/client/components/router-reducer/router-reducer.ts
@@ -2,516 +2,20 @@ import { CacheNode, CacheStates } from '../../../shared/lib/app-router-context'
 import type {
   FlightRouterState,
   FlightData,
-  FlightDataPath,
   FlightSegmentPath,
-  Segment,
 } from '../../../server/app-render'
-import { matchSegment } from '../match-segments'
 import { fetchServerResponse } from '../app-router'
 import { createRecordFromThenable } from './create-record-from-thenable'
 import { readRecordValue } from './read-record-value'
 import { createHrefFromUrl } from './create-href-from-url'
-
-/**
- * Invalidate cache one level down from the router state.
- */
-function invalidateCacheByRouterState(
-  newCache: CacheNode,
-  existingCache: CacheNode,
-  routerState: FlightRouterState
-): void {
-  // Remove segment that we got data for so that it is filled in during rendering of subTreeData.
-  for (const key in routerState[1]) {
-    const segmentForParallelRoute = routerState[1][key][0]
-    const cacheKey = Array.isArray(segmentForParallelRoute)
-      ? segmentForParallelRoute[1]
-      : segmentForParallelRoute
-    const existingParallelRoutesCacheNode =
-      existingCache.parallelRoutes.get(key)
-    if (existingParallelRoutesCacheNode) {
-      let parallelRouteCacheNode = new Map(existingParallelRoutesCacheNode)
-      parallelRouteCacheNode.delete(cacheKey)
-      newCache.parallelRoutes.set(key, parallelRouteCacheNode)
-    }
-  }
-}
-
-function fillLazyItemsTillLeafWithHead(
-  newCache: CacheNode,
-  existingCache: CacheNode | undefined,
-  routerState: FlightRouterState,
-  head: React.ReactNode
-): void {
-  const isLastSegment = Object.keys(routerState[1]).length === 0
-  if (isLastSegment) {
-    newCache.head = head
-    return
-  }
-  // Remove segment that we got data for so that it is filled in during rendering of subTreeData.
-  for (const key in routerState[1]) {
-    const parallelRouteState = routerState[1][key]
-    const segmentForParallelRoute = parallelRouteState[0]
-    const cacheKey = Array.isArray(segmentForParallelRoute)
-      ? segmentForParallelRoute[1]
-      : segmentForParallelRoute
-    if (existingCache) {
-      const existingParallelRoutesCacheNode =
-        existingCache.parallelRoutes.get(key)
-      if (existingParallelRoutesCacheNode) {
-        let parallelRouteCacheNode = new Map(existingParallelRoutesCacheNode)
-        parallelRouteCacheNode.delete(cacheKey)
-        const newCacheNode: CacheNode = {
-          status: CacheStates.LAZY_INITIALIZED,
-          data: null,
-          subTreeData: null,
-          parallelRoutes: new Map(),
-        }
-        parallelRouteCacheNode.set(cacheKey, newCacheNode)
-        fillLazyItemsTillLeafWithHead(
-          newCacheNode,
-          undefined,
-          parallelRouteState,
-          head
-        )
-
-        newCache.parallelRoutes.set(key, parallelRouteCacheNode)
-        continue
-      }
-    }
-
-    const newCacheNode: CacheNode = {
-      status: CacheStates.LAZY_INITIALIZED,
-      data: null,
-      subTreeData: null,
-      parallelRoutes: new Map(),
-    }
-    newCache.parallelRoutes.set(key, new Map([[cacheKey, newCacheNode]]))
-    fillLazyItemsTillLeafWithHead(
-      newCacheNode,
-      undefined,
-      parallelRouteState,
-      head
-    )
-  }
-}
-
-/**
- * Fill cache with subTreeData based on flightDataPath
- */
-function fillCacheWithNewSubTreeData(
-  newCache: CacheNode,
-  existingCache: CacheNode,
-  flightDataPath: FlightDataPath
-): void {
-  const isLastEntry = flightDataPath.length <= 5
-  const [parallelRouteKey, segment] = flightDataPath
-
-  const segmentForCache = Array.isArray(segment) ? segment[1] : segment
-
-  const existingChildSegmentMap =
-    existingCache.parallelRoutes.get(parallelRouteKey)
-
-  if (!existingChildSegmentMap) {
-    // Bailout because the existing cache does not have the path to the leaf node
-    // Will trigger lazy fetch in layout-router because of missing segment
-    return
-  }
-
-  let childSegmentMap = newCache.parallelRoutes.get(parallelRouteKey)
-  if (!childSegmentMap || childSegmentMap === existingChildSegmentMap) {
-    childSegmentMap = new Map(existingChildSegmentMap)
-    newCache.parallelRoutes.set(parallelRouteKey, childSegmentMap)
-  }
-
-  const existingChildCacheNode = existingChildSegmentMap.get(segmentForCache)
-  let childCacheNode = childSegmentMap.get(segmentForCache)
-
-  if (isLastEntry) {
-    if (
-      !childCacheNode ||
-      !childCacheNode.data ||
-      childCacheNode === existingChildCacheNode
-    ) {
-      childCacheNode = {
-        status: CacheStates.READY,
-        data: null,
-        subTreeData: flightDataPath[3],
-        // Ensure segments other than the one we got data for are preserved.
-        parallelRoutes: existingChildCacheNode
-          ? new Map(existingChildCacheNode.parallelRoutes)
-          : new Map(),
-      }
-
-      if (existingChildCacheNode) {
-        invalidateCacheByRouterState(
-          childCacheNode,
-          existingChildCacheNode,
-          flightDataPath[2]
-        )
-      }
-
-      fillLazyItemsTillLeafWithHead(
-        childCacheNode,
-        existingChildCacheNode,
-        flightDataPath[2],
-        flightDataPath[4]
-      )
-
-      childSegmentMap.set(segmentForCache, childCacheNode)
-    }
-    return
-  }
-
-  if (!childCacheNode || !existingChildCacheNode) {
-    // Bailout because the existing cache does not have the path to the leaf node
-    // Will trigger lazy fetch in layout-router because of missing segment
-    return
-  }
-
-  if (childCacheNode === existingChildCacheNode) {
-    childCacheNode = {
-      status: childCacheNode.status,
-      data: childCacheNode.data,
-      subTreeData: childCacheNode.subTreeData,
-      parallelRoutes: new Map(childCacheNode.parallelRoutes),
-    } as CacheNode
-    childSegmentMap.set(segmentForCache, childCacheNode)
-  }
-
-  fillCacheWithNewSubTreeData(
-    childCacheNode,
-    existingChildCacheNode,
-    flightDataPath.slice(2)
-  )
-}
-
-/**
- * Fill cache up to the end of the flightSegmentPath, invalidating anything below it.
- */
-function invalidateCacheBelowFlightSegmentPath(
-  newCache: CacheNode,
-  existingCache: CacheNode,
-  flightSegmentPath: FlightSegmentPath
-): void {
-  const isLastEntry = flightSegmentPath.length <= 2
-  const [parallelRouteKey, segment] = flightSegmentPath
-
-  const segmentForCache = Array.isArray(segment) ? segment[1] : segment
-
-  const existingChildSegmentMap =
-    existingCache.parallelRoutes.get(parallelRouteKey)
-
-  if (!existingChildSegmentMap) {
-    // Bailout because the existing cache does not have the path to the leaf node
-    // Will trigger lazy fetch in layout-router because of missing segment
-    return
-  }
-
-  let childSegmentMap = newCache.parallelRoutes.get(parallelRouteKey)
-  if (!childSegmentMap || childSegmentMap === existingChildSegmentMap) {
-    childSegmentMap = new Map(existingChildSegmentMap)
-    newCache.parallelRoutes.set(parallelRouteKey, childSegmentMap)
-  }
-
-  // In case of last entry don't copy further down.
-  if (isLastEntry) {
-    childSegmentMap.delete(segmentForCache)
-    return
-  }
-
-  const existingChildCacheNode = existingChildSegmentMap.get(segmentForCache)
-  let childCacheNode = childSegmentMap.get(segmentForCache)
-
-  if (!childCacheNode || !existingChildCacheNode) {
-    // Bailout because the existing cache does not have the path to the leaf node
-    // Will trigger lazy fetch in layout-router because of missing segment
-    return
-  }
-
-  if (childCacheNode === existingChildCacheNode) {
-    childCacheNode = {
-      status: childCacheNode.status,
-      data: childCacheNode.data,
-      subTreeData: childCacheNode.subTreeData,
-      parallelRoutes: new Map(childCacheNode.parallelRoutes),
-    } as CacheNode
-    childSegmentMap.set(segmentForCache, childCacheNode)
-  }
-
-  invalidateCacheBelowFlightSegmentPath(
-    childCacheNode,
-    existingChildCacheNode,
-    flightSegmentPath.slice(2)
-  )
-}
-
-/**
- * Kick off fetch based on the common layout between two routes. Fill cache with data property holding the in-progress fetch.
- */
-function fillCacheWithDataProperty(
-  newCache: CacheNode,
-  existingCache: CacheNode,
-  segments: string[],
-  fetchResponse: () => ReturnType<typeof fetchServerResponse>
-): { bailOptimistic: boolean } | undefined {
-  const isLastEntry = segments.length === 1
-
-  const parallelRouteKey = 'children'
-  const [segment] = segments
-
-  const existingChildSegmentMap =
-    existingCache.parallelRoutes.get(parallelRouteKey)
-
-  if (!existingChildSegmentMap) {
-    // Bailout because the existing cache does not have the path to the leaf node
-    // Will trigger lazy fetch in layout-router because of missing segment
-    return { bailOptimistic: true }
-  }
-
-  let childSegmentMap = newCache.parallelRoutes.get(parallelRouteKey)
-
-  if (!childSegmentMap || childSegmentMap === existingChildSegmentMap) {
-    childSegmentMap = new Map(existingChildSegmentMap)
-    newCache.parallelRoutes.set(parallelRouteKey, childSegmentMap)
-  }
-
-  const existingChildCacheNode = existingChildSegmentMap.get(segment)
-  let childCacheNode = childSegmentMap.get(segment)
-
-  // In case of last segment start off the fetch at this level and don't copy further down.
-  if (isLastEntry) {
-    if (
-      !childCacheNode ||
-      !childCacheNode.data ||
-      childCacheNode === existingChildCacheNode
-    ) {
-      childSegmentMap.set(segment, {
-        status: CacheStates.DATA_FETCH,
-        data: fetchResponse(),
-        subTreeData: null,
-        parallelRoutes: new Map(),
-      })
-    }
-    return
-  }
-
-  if (!childCacheNode || !existingChildCacheNode) {
-    // Start fetch in the place where the existing cache doesn't have the data yet.
-    if (!childCacheNode) {
-      childSegmentMap.set(segment, {
-        status: CacheStates.DATA_FETCH,
-        data: fetchResponse(),
-        subTreeData: null,
-        parallelRoutes: new Map(),
-      })
-    }
-    return
-  }
-
-  if (childCacheNode === existingChildCacheNode) {
-    childCacheNode = {
-      status: childCacheNode.status,
-      data: childCacheNode.data,
-      subTreeData: childCacheNode.subTreeData,
-      parallelRoutes: new Map(childCacheNode.parallelRoutes),
-    } as CacheNode
-    childSegmentMap.set(segment, childCacheNode)
-  }
-
-  return fillCacheWithDataProperty(
-    childCacheNode,
-    existingChildCacheNode,
-    segments.slice(1),
-    fetchResponse
-  )
-}
-
-/**
- * Create optimistic version of router state based on the existing router state and segments.
- * This is used to allow rendering layout-routers up till the point where data is missing.
- */
-function createOptimisticTree(
-  segments: string[],
-  flightRouterState: FlightRouterState | null,
-  _isFirstSegment: boolean,
-  parentRefetch: boolean,
-  _href?: string
-): FlightRouterState {
-  const [existingSegment, existingParallelRoutes] = flightRouterState || [
-    null,
-    {},
-  ]
-  const segment = segments[0]
-  const isLastSegment = segments.length === 1
-
-  const segmentMatches =
-    existingSegment !== null && matchSegment(existingSegment, segment)
-  const shouldRefetchThisLevel = !flightRouterState || !segmentMatches
-
-  let parallelRoutes: FlightRouterState[1] = {}
-  if (existingSegment !== null && segmentMatches) {
-    parallelRoutes = existingParallelRoutes
-  }
-
-  let childTree
-  if (!isLastSegment) {
-    const childItem = createOptimisticTree(
-      segments.slice(1),
-      parallelRoutes ? parallelRoutes.children : null,
-      false,
-      parentRefetch || shouldRefetchThisLevel
-    )
-
-    childTree = childItem
-  }
-
-  const result: FlightRouterState = [
-    segment,
-    {
-      ...parallelRoutes,
-      ...(childTree ? { children: childTree } : {}),
-    },
-  ]
-
-  if (!parentRefetch && shouldRefetchThisLevel) {
-    result[3] = 'refetch'
-  }
-
-  return result
-}
-
-/**
- * Apply the router state from the Flight response. Creates a new router state tree.
- */
-function applyRouterStatePatchToTree(
-  flightSegmentPath: FlightData[0],
-  flightRouterState: FlightRouterState,
-  treePatch: FlightRouterState
-): FlightRouterState | null {
-  const [segment, parallelRoutes, , , isRootLayout] = flightRouterState
-
-  // Root refresh
-  if (flightSegmentPath.length === 1) {
-    const tree: FlightRouterState = [...treePatch]
-
-    return tree
-  }
-
-  const [currentSegment, parallelRouteKey] = flightSegmentPath
-
-  // Tree path returned from the server should always match up with the current tree in the browser
-  if (!matchSegment(currentSegment, segment)) {
-    return null
-  }
-
-  const lastSegment = flightSegmentPath.length === 2
-
-  let parallelRoutePatch
-  if (lastSegment) {
-    parallelRoutePatch = treePatch
-  } else {
-    parallelRoutePatch = applyRouterStatePatchToTree(
-      flightSegmentPath.slice(2),
-      parallelRoutes[parallelRouteKey],
-      treePatch
-    )
-
-    if (parallelRoutePatch === null) {
-      return null
-    }
-  }
-
-  const tree: FlightRouterState = [
-    flightSegmentPath[0],
-    {
-      ...parallelRoutes,
-      [parallelRouteKey]: parallelRoutePatch,
-    },
-  ]
-
-  // Current segment is the root layout
-  if (isRootLayout) {
-    tree[4] = true
-  }
-
-  return tree
-}
-
-function shouldHardNavigate(
-  flightSegmentPath: FlightDataPath,
-  flightRouterState: FlightRouterState,
-  treePatch: FlightRouterState
-): boolean {
-  const [segment, parallelRoutes] = flightRouterState
-  // TODO-APP: Check if `as` can be replaced.
-  const [currentSegment, parallelRouteKey] = flightSegmentPath as [
-    Segment,
-    string
-  ]
-
-  // Check if current segment matches the existing segment.
-  if (!matchSegment(currentSegment, segment)) {
-    // If dynamic parameter in tree doesn't match up with segment path a hard navigation is triggered.
-    if (Array.isArray(currentSegment)) {
-      return true
-    }
-
-    // If the existing segment did not match soft navigation is triggered.
-    return false
-  }
-  const lastSegment = flightSegmentPath.length <= 2
-
-  if (lastSegment) {
-    return false
-  }
-
-  return shouldHardNavigate(
-    flightSegmentPath.slice(2),
-    parallelRoutes[parallelRouteKey],
-    treePatch
-  )
-}
-
-function isNavigatingToNewRootLayout(
-  currentTree: FlightRouterState,
-  nextTree: FlightRouterState
-): boolean {
-  // Compare segments
-  const currentTreeSegment = currentTree[0]
-  const nextTreeSegment = nextTree[0]
-  // If any segment is different before we find the root layout, the root layout has changed.
-  // E.g. /same/(group1)/layout.js -> /same/(group2)/layout.js
-  // First segment is 'same' for both, keep looking. (group1) changed to (group2) before the root layout was found, it must have changed.
-  if (Array.isArray(currentTreeSegment) && Array.isArray(nextTreeSegment)) {
-    // Compare dynamic param name and type but ignore the value, different values would not affect the current root layout
-    // /[name] - /slug1 and /slug2, both values (slug1 & slug2) still has the same layout /[name]/layout.js
-    if (
-      currentTreeSegment[0] !== nextTreeSegment[0] ||
-      currentTreeSegment[2] !== nextTreeSegment[2]
-    ) {
-      return true
-    }
-  } else if (currentTreeSegment !== nextTreeSegment) {
-    return true
-  }
-
-  // Current tree root layout found
-  if (currentTree[4]) {
-    // If the next tree doesn't have the root layout flag, it must have changed.
-    return !nextTree[4]
-  }
-  // Current tree  didn't have its root layout here, must have changed.
-  if (nextTree[4]) {
-    return true
-  }
-  // We can't assume it's `parallelRoutes.children` here in case the root layout is `app/@something/layout.js`
-  // But it's not possible to be more than one parallelRoutes before the root layout is found
-  // TODO-APP: change to traverse all parallel routes
-  const currentTreeChild = Object.values(currentTree[1])[0]
-  const nextTreeChild = Object.values(nextTree[1])[0]
-  if (!currentTreeChild || !nextTreeChild) return true
-  return isNavigatingToNewRootLayout(currentTreeChild, nextTreeChild)
-}
+import { fillLazyItemsTillLeafWithHead } from './fill-lazy-items-till-leaf-with-head'
+import { fillCacheWithNewSubTreeData } from './fill-cache-with-new-sub-tree-data'
+import { invalidateCacheBelowFlightSegmentPath } from './invalidateCacheBelowFlightSegmentPath'
+import { fillCacheWithDataProperty } from './fill-cache-with-data-property'
+import { createOptimisticTree } from './create-optimistic-tree'
+import { applyRouterStatePatchToTree } from './apply-router-state-patch-to-tree'
+import { shouldHardNavigate } from './should-hard-navigate'
+import { isNavigatingToNewRootLayout } from './is-navigating-to-new-root-layout'
 
 export type FocusAndScrollRef = {
   /**
